+++ src/size/size.ts
@@ -12,6 +12,8 @@ interface RestorableThemes {
   themes: Set<Theme>
 }
 
+const sizePluginSymbol = Symbol('marp-size-plugin')
+
 export const markdown = marpitPlugin(md => {
   const marp: Marp = md.marpit
   const { render } = marp
@@ -38,9 +40,10 @@ export const markdown = marpitPlugin(md => {
     default: undefined,
   }
 
-  // `size` global directive
-  marp.customDirectives.global.size = size =>
-    typeof size === 'string' ? { size } : {}
+  // Define `size` global directive
+  Object.defineProperty(marp.customDirectives.global, 'size', {
+    value: size => (typeof size === 'string' ? { size } : {}),
+  })
 
   // Override render method to restore original theme set
   marp.render = (...args) => {
@@ -66,6 +69,8 @@ export const markdown = marpitPlugin(md => {
     const customSize = definedSizes(themeInstance).get(size)
 
     if (customSize) {
+      state[sizePluginSymbol] = size
+
       const { width, height } = customSize
       const css = `${themeInstance.css}\nsection{width:${width};height:${height};}`
 
@@ -87,4 +92,27 @@ export const markdown = marpitPlugin(md => {
       }
     }
   })
+
+  md.core.ruler.after('marpit_directives_apply', 'marp_size_apply', state => {
+    if (state.inlineMode || !state[sizePluginSymbol]) return
+
+    for (const token of state.tokens) {
+      const { marpitDirectives } = token.meta || {}
+      if (marpitDirectives) token.attrSet('data-size', state[sizePluginSymbol])
+    }
+  })
+
+  md.core.ruler.after(
+    'marpit_advanced_background',
+    'marp_size_apply_advanced_background',
+    state => {
+      if (state.inlineMode || !state[sizePluginSymbol]) return
+
+      for (const token of state.tokens) {
+        if (token.type === 'marpit_advanced_pseudo_section_open') {
+          token.attrSet('data-size', state[sizePluginSymbol])
+        }
+      }
+    }
+  )
 })